feat(types): manually create type declaration files within /declarations
